use optparse and move calc to main function
