+++ main.cc
@@ -1,48 +1,55 @@
 #include <iostream>
 #include <string>
-#include <vector>
-#include <regex>
+#include <optparse/parser.hpp>
 #include <cmath>
-#include "algrithms/blobrating.cc"
 
-static std::string *commandLineArgs = new std::string {""};
-static std::vector<std::string> *options = new std::vector<std::string> {};
-
-bool vectorIncludes(std::vector<std::string> &vec, std::string key) {
-	bool includes = false;
-	for(std::string& v : vec) {
-		if (v == key) includes = true;
-	}
-	return includes;
+short calc(unsigned int own_br, unsigned int br_of_opponent)
+{
+        int gainValue = 0;
+        int br_difference = own_br - br_of_opponent;
+        if (br_difference < 0)
+        {
+                gainValue = pow(log10(-br_difference), 4);
+        }
+        else
+        {
+                gainValue = log1p(9998 / fdim(own_br, br_of_opponent));
+        }
+        if (own_br + gainValue >= 9999)
+        {
+                return 0;
+        }
+        if (gainValue > 999 || gainValue < 0) return 32;
+        return gainValue == 0 ? ++gainValue : gainValue;
 }
 
 int main(int argc, char **argv)
 {
-	// argv to string
-	for (int i = 0; i < argc; ++i) {
-		*commandLineArgs += " ";
-		*commandLineArgs += argv[i];
+    // Options
+    // m = own BR
+    // o = opponent BR
+    optparse::parser parser(argc, argv, "mo");
+	optparse::option* opts = parser.get_options();
+
+	int own_br,
+	opponent_br;
+
+	for (int i = 0;; ++i) {
+	    if (!opts[i].name) break;
+
+        try {
+	        if (opts[i].name == 'm') {
+	            own_br = std::stoi(opts[i].value);
+	        } else if (opts[i].name == 'o') {
+	            opponent_br = std::stoi(opts[i].value);
+	        }
+	    } catch(...) {
+	        std::cerr << "Invalid BR value\n";
+	    }
 	}
-	*commandLineArgs = commandLineArgs->substr(1);
-	
-	// Parse options
-	std::smatch matched;
-	std::regex optionExpr("--\\w+");
-	std::regex_search(*commandLineArgs, matched, optionExpr);
-	
-	for (int i = 0; i < matched.size(); ++i) {
-		options->push_back(matched[i]);
-	}
-	
-	if(vectorIncludes(*options, "--br")) {
-		unsigned int ownBR = atoi(argv[1]);
-		unsigned int opponentBR = atoi(argv[2]);
-		std::cout << calc(ownBR, opponentBR);
-	}
-	
-	
-	delete commandLineArgs;
-	delete options;
+
+	std::cout << calc(own_br, opponent_br);
+
 	return 0;
 }
